Add 2.7 phonebook
There's an issue with adding a contact list. When adding a duplicate name,
the alert pops up and resets the input. But when the same name is added again,
alert pops up but the duplicate name is added to the list. This is a bug
and will need to be fixed on later date.